feat: Add display order to tags and categories
